+++ impl/src/main/java/edu/si/trellis/query/rdf/ImmutableRetrieve.java
@@ -1,6 +1,10 @@
 package edu.si.trellis.query.rdf;
 
+import static java.util.stream.StreamSupport.stream;
+
 import com.datastax.driver.core.ConsistencyLevel;
+import com.datastax.driver.core.ResultSet;
+import com.datastax.driver.core.Row;
 import com.datastax.driver.core.Session;
 
 import edu.si.trellis.MutableReadConsistency;
@@ -10,6 +14,7 @@ import java.util.stream.Stream;
 
 import javax.inject.Inject;
 
+import org.apache.commons.rdf.api.Dataset;
 import org.apache.commons.rdf.api.IRI;
 import org.apache.commons.rdf.api.Quad;
 
@@ -28,6 +33,14 @@ public class ImmutableRetrieve extends ResourceQuery {
      * @return the RDF retrieved
      */
     public CompletionStage<Stream<Quad>> execute(IRI id) {
-        return quads(preparedStatement().bind().set("identifier", id, IRI.class));
+        return executeRead(preparedStatement().bind().set("identifier", id, IRI.class))
+                        .thenApply(ResultSet::spliterator)
+                        .thenApply(r -> stream(r, false))
+                        .thenApply(row -> row.map(this::getDataset))
+                        .thenApply(r -> r.flatMap(Dataset::stream));
+    }
+
+    private Dataset getDataset(Row r) {
+        return r.get("quads", Dataset.class);
     }
 }
